+++ DependencyInjection/Configuration.php
@@ -15,49 +15,63 @@ use Symfony\Component\Config\Definition\ConfigurationInterface;
  */
 class Configuration implements ConfigurationInterface
 {
+
     /**
      * {@inheritDoc}
      */
     public function getConfigTreeBuilder()
     {
         $treeBuilder = new TreeBuilder();
-        $rootNode = $treeBuilder->root('fr3d_ldap');
+        $rootNode    = $treeBuilder->root('fr3d_ldap');
 
         $rootNode
-            ->children()
-              ->arrayNode('client')
                 ->children()
-                  ->scalarNode('host')->isRequired()->cannotBeEmpty()->end()
-                  ->scalarNode('port')->defaultValue(389)->end()
-                  ->scalarNode('version')->defaultValue(3)->end()
-                  ->scalarNode('useSsl')->defaultFalse()->end()
-                  ->scalarNode('useStartTls')->defaultFalse()->end()
-                  ->scalarNode('username')->end()
-                  ->scalarNode('password')->end()
-                  ->scalarNode('optReferrals')->end()
-                ->end()
-              ->end()
-              ->arrayNode('user')
-                ->children()
-                  ->scalarNode('baseDn')->isRequired()->cannotBeEmpty()->end()
-                  ->scalarNode('filter')->defaultValue('')->end()
-                  ->arrayNode('attributes')
-                    ->addDefaultsIfNotSet()
-                    ->defaultValue(array(array('ldap_attr' => 'uid', 'user_method' => 'setUsername')))
-                    ->prototype('array')
-                      ->children()
-                        ->scalarNode('ldap_attr')->isRequired()->cannotBeEmpty()->end()
-                        ->scalarNode('user_method')->isRequired()->cannotBeEmpty()->end()
-                      ->end()
+                    ->arrayNode('driver')
+                        ->children()
+                            ->scalarNode('host')->isRequired()->cannotBeEmpty()->end()
+                            ->scalarNode('port')->defaultValue(389)->end()
+                            ->scalarNode('useStartTls')->defaultFalse()->end()
+                            ->scalarNode('useSsl')->defaultFalse()->end()
+                            ->scalarNode('username')->end()
+                            ->scalarNode('password')->end()
+                            ->scalarNode('bindRequiresDn')->defaultFalse()->end()
+                            ->scalarNode('baseDn')->end()
+                            ->scalarNode('accountCanonicalForm')->end()
+                            ->scalarNode('accountDomainName')->end()
+                            ->scalarNode('accountDomainNameShort')->end()
+                            ->scalarNode('accountFilterFormat')->end()
+                            ->scalarNode('allowEmptyPassword')->end()
+                            ->scalarNode('optReferrals')->end()
+                            ->scalarNode('tryUsernameSplit')->end()
+                        ->end()
                     ->end()
-                  ->end()
+                    ->arrayNode('user')
+                        ->children()
+                            ->scalarNode('baseDn')->isRequired()->cannotBeEmpty()->end()
+                            ->scalarNode('filter')->defaultValue('')->end()
+                            ->arrayNode('attributes')
+                                ->addDefaultsIfNotSet()
+                                ->defaultValue(array(
+                                    array(
+                                        'ldap_attr'   => 'uid',
+                                        'user_method' => 'setUsername')
+                                    ))
+                                ->prototype('array')
+                                    ->children()
+                                        ->scalarNode('ldap_attr')->isRequired()->cannotBeEmpty()->end()
+                                        ->scalarNode('user_method')->isRequired()->cannotBeEmpty()->end()
+                                    ->end()
+                                ->end()
+                            ->end()
+                        ->end()
+                    ->end()
                 ->end()
-              ->end()
-            ->end()
-            ->validate()
-              ->ifTrue(function($v){return $v['client']['useSsl'] && $v['client']['useStartTls'];})
-              ->thenInvalid('The useSsl and useStartTls options are mutually exclusive.')
-            ->end();
+                ->validate()
+                    ->ifTrue(function($v) {
+                                return $v['driver']['useSsl'] && $v['driver']['useStartTls'];
+                            })
+                    ->thenInvalid('The useSsl and useStartTls options are mutually exclusive.')
+                ->end();
 
         $this->addServiceSection($rootNode);
 
@@ -67,16 +81,17 @@ class Configuration implements ConfigurationInterface
     private function addServiceSection(ArrayNodeDefinition $node)
     {
         $node
-            ->addDefaultsIfNotSet()
-            ->children()
-              ->arrayNode('service')
                 ->addDefaultsIfNotSet()
-                  ->children()
-                    ->scalarNode('user_manager')->defaultValue('fos_user.user_manager')->end()
-                    ->scalarNode('ldap_manager')->defaultValue('fr3d_ldap.ldap_manager.default')->end()
-                  ->end()
-                ->end()
-              ->end()
-            ->end();
+                    ->children()
+                        ->arrayNode('service')
+                            ->addDefaultsIfNotSet()
+                            ->children()
+                                ->scalarNode('user_manager')->defaultValue('fos_user.user_manager')->end()
+                                ->scalarNode('ldap_manager')->defaultValue('fr3d_ldap.ldap_manager.default')->end()
+                                ->scalarNode('ldap_driver')->defaultValue('fr3d_ldap.ldap_driver.zend')->end()
+                            ->end()
+                        ->end()
+                    ->end()
+                ->end();
     }
 }