+++ DependencyInjection/FR3DLdapExtension.php
@@ -19,14 +19,22 @@ class FR3DLdapExtension extends Extension
         $config        = $this->processConfiguration($configuration, $configs);
 
         $loader = new Loader\XmlFileLoader($container, new FileLocator(__DIR__ . '/../Resources/config'));
-        foreach (array('services', 'validator') as $basename) {
+        foreach (array('services', 'security', 'validator', 'ldap_driver') as $basename) {
             $loader->load(sprintf('%s.xml', $basename));
         }
 
         $container->setAlias('fr3d_ldap.user_manager', $config['service']['user_manager']);
         $container->setAlias('fr3d_ldap.ldap_manager', $config['service']['ldap_manager']);
+        $container->setAlias('fr3d_ldap.ldap_driver', $config['service']['ldap_driver']);
 
-        $container->setParameter('fr3d_ldap.client.parameters', $config['client']);
+        if (!isset($config['driver']['baseDn'])) {
+            $config['driver']['baseDn'] = $config['user']['baseDn'];
+        }
+        if (!isset($config['driver']['accountFilterFormat'])) {
+            $config['driver']['accountFilterFormat'] = $config['user']['filter'];
+        }
+
+        $container->setParameter('fr3d_ldap.ldap_driver.parameters', $config['driver']);
         $container->setParameter('fr3d_ldap.ldap_manager.parameters', $config['user']);
     }