+++ Validator/UniqueValidator.php
@@ -56,9 +56,9 @@ class UniqueValidator extends ConstraintValidator
         $user = $this->ldapManager->findUserByUsername($value->getUsername());
 
         if ($user) {
-            $this->setMessage($constraint->message, array(
-                '%property%' => $constraint->property
-            ));
+            $this->context->addViolation($constraint->message, array(
+                '%property%' => $constraint->property,
+            ), $value);
             return false;
         }